Change to consider "gaagaag" exception
The "gaagaag" exception consists in those cases when a repeated char (in this case 'g') also plays the role
of an isolated char that can moved to the center of the string to form palindromes with odd-sized strings.
This is different from a case like "gaataag", where the isolated char is no-repeated char.

These changes make possible to this app to consider this exception in its logic.