changed build tree algorithm
PruneStrategy changes its way to build a binary tree from the given missile
data. Before, it was appending child nodes directly to the tree's root node.
Now, it's appending a child node to the last appended node.

e.g. the binary tree for [5000, 6000, 5500, 5000, 45000] with the previous
algorithm is this:
                                   --------
                -------------------| 5000 |-------------------
		|	           --------                  |
		|                                            |
		|					     |
             --------				          --------
     --------| 5000 |                        -------------| 6000 |
     |       --------                        |            --------
     |                                       |
     |	                                     |
 --------                                 ---------
 | 4500 |				  | 5500  |
 --------                                 ---------

The binary tree for the same missile data with the new algorithm is this:
                                 --------
				 | 5000 |--------------------------
				 --------                         |
				                                  |
                                                                  |
						               --------
			                            -----------| 6000 |
					            |	       --------
                                                    |
                                                    |
                                                 --------
		                      -----------| 5500 |
				      |	         --------
				      |
				      |
                                  --------
                       -----------| 5000 |
		       |	  --------
                       |
                       |
                    --------
		    | 4500 |
                    --------

The first algorithm is going to tell wrong the optimal missile destruction route
in some cases, like the one represented in the above trees. In this specific case,
the optimal destruction route (the one that destroys most missiles) is [2, 3, 4, 5],
but this route isn't present in the first tree. This makes impossible for the first
algorithm to tell the right answer in cases like this. The second tree does have
that optimal destruction route. It can tell the right answer for cases like this,
among other kinds of cases. That's the reason behind these changes.